Calculates best case and worst case for emission profile at product level

Calculates best case and worst case for emission profile at product level

Usage

best_case_worst_case_emission_profile(data)

Arguments

data

Dataframe. Emissions profile product level output

Value

A dataframe

Examples

product <- unnest_product(toy_profile_emissions_impl_output())

product |> best_case_worst_case_emission_profile()
#> # A tibble: 456 × 36
#>    companies_id       company_name country emission_profile benchmark ep_product
#>    <chr>              <chr>        <chr>   <chr>            <chr>     <chr>     
#>  1 asteria_megalotom… asteria_meg… austria high             all       tent      
#>  2 asteria_megalotom… asteria_meg… austria high             isic_4di… tent      
#>  3 asteria_megalotom… asteria_meg… austria high             tilt_sub… tent      
#>  4 asteria_megalotom… asteria_meg… austria high             unit      tent      
#>  5 asteria_megalotom… asteria_meg… austria high             unit_isi… tent      
#>  6 asteria_megalotom… asteria_meg… austria high             unit_til… tent      
#>  7 skarn_gallinule    skarn_galli… austria high             all       sheds, co…
#>  8 skarn_gallinule    skarn_galli… austria high             isic_4di… sheds, co…
#>  9 skarn_gallinule    skarn_galli… austria high             tilt_sub… sheds, co…
#> 10 skarn_gallinule    skarn_galli… austria high             unit      sheds, co…
#> # ℹ 446 more rows
#> # ℹ 30 more variables: matched_activity_name <chr>,
#> #   matched_reference_product <chr>, unit <chr>, multi_match <lgl>,
#> #   matching_certainty <chr>, matching_certainty_company_average <chr>,
#> #   tilt_sector <chr>, tilt_subsector <chr>, isic_4digit <chr>,
#> #   isic_4digit_name <chr>, company_city <chr>, postcode <chr>, address <chr>,
#> #   main_activity <chr>, activity_uuid_product_uuid <chr>, …