Calculates best case and worst case for sector profile at product level

Calculates best case and worst case for sector profile at product level

Usage

best_case_worst_case_sector_profile(data)

Arguments

data

Dataframe. Sector profile nested output

Value

A dataframe

Examples

out <- toy_profile_sector_impl_output()

out |> best_case_worst_case_sector_profile()
#> # A tibble: 72 × 3
#>    companies_id                       product           company           
#>  * <chr>                              <list>            <list>            
#>  1 antimonarchy_canine                <tibble [4 × 32]> <tibble [16 × 12]>
#>  2 celestial_lovebird                 <tibble [4 × 32]> <tibble [16 × 12]>
#>  3 nonphilosophical_llama             <tibble [8 × 32]> <tibble [16 × 12]>
#>  4 asteria_megalotomusquinquespinosus <tibble [4 × 32]> <tibble [16 × 12]>
#>  5 quasifaithful_amphiuma             <tibble [4 × 32]> <tibble [16 × 12]>
#>  6 spectacular_americanriverotter     <tibble [4 × 32]> <tibble [16 × 12]>
#>  7 contrite_silkworm                  <tibble [4 × 32]> <tibble [16 × 12]>
#>  8 harmless_owlbutterfly              <tibble [4 × 32]> <tibble [16 × 12]>
#>  9 fascist_maiasaura                  <tibble [4 × 32]> <tibble [16 × 12]>
#> 10 charismatic_islandwhistler         <tibble [4 × 32]> <tibble [16 × 12]>
#> # ℹ 62 more rows