Calculate the indicator "transition risk profile"

Adds the risk classification to calculated transition risk scores from emission profile and sector profile indicator.

Usage

transition_risk_profile(
  emissions_profile,
  sector_profile,
  co2,
  all_activities_scenario_sectors,
  scenarios,
  pivot_wider = FALSE
)

Arguments

emissions_profile

Nested data frame. The output of profile_emissions().

sector_profile

Nested data frame. The output of profile_sector().

co2

A dataframe

all_activities_scenario_sectors

A dataframe

scenarios

A dataframe

pivot_wider

Logical. Pivot the output at company level to a wide format?

Value

A data frame with the column companies_id, and the nested columnsproduct and company holding the outputs at product and company level.

See also

Examples

library(readr, warn.conflicts = FALSE)
library(dplyr, warn.conflicts = FALSE)
library(tiltToyData, warn.conflicts = FALSE)
library(tiltTransitionRisk, warn.conflicts = FALSE)

set.seed(123)
restore <- options(list(
  readr.show_col_types = FALSE,
  tiltIndicatorAfter.output_co2_footprint = TRUE
))

toy_emissions_profile_products_ecoinvent <- read_csv(toy_emissions_profile_products_ecoinvent())
toy_emissions_profile_any_companies <- read_csv(toy_emissions_profile_any_companies())
toy_sector_profile_any_scenarios <- read_csv(toy_sector_profile_any_scenarios())
toy_sector_profile_companies <- read_csv(toy_sector_profile_companies())
toy_europages_companies <- read_csv(toy_europages_companies())
toy_ecoinvent_activities <- read_csv(toy_ecoinvent_activities())
toy_ecoinvent_europages <- read_csv(toy_ecoinvent_europages())
toy_ecoinvent_inputs <- read_csv(toy_ecoinvent_inputs())
toy_isic_name <- read_csv(toy_isic_name())
toy_all_activities_scenario_sectors <- read_csv(toy_all_activities_scenario_sectors())

toy_emissions_profile <- profile_emissions(
  companies = toy_emissions_profile_any_companies,
  co2 = toy_emissions_profile_products_ecoinvent,
  europages_companies = toy_europages_companies,
  ecoinvent_activities = toy_ecoinvent_activities,
  ecoinvent_europages = toy_ecoinvent_europages,
  isic = toy_isic_name
)

toy_sector_profile <- profile_sector(
  companies = toy_sector_profile_companies,
  scenarios = toy_sector_profile_any_scenarios,
  europages_companies = toy_europages_companies,
  ecoinvent_activities = toy_ecoinvent_activities,
  ecoinvent_europages = toy_ecoinvent_europages,
  isic = toy_isic_name
)

output <- transition_risk_profile(
  emissions_profile = toy_emissions_profile,
  sector_profile = toy_sector_profile,
  co2 = toy_emissions_profile_products_ecoinvent,
  all_activities_scenario_sectors = toy_all_activities_scenario_sectors,
  scenarios = toy_sector_profile_any_scenarios,
  pivot_wider = FALSE
)

output |> unnest_product()
#> # A tibble: 1,824 × 28
#>    companies_id                     activity_uuid_produc…¹ transition_risk_score
#>    <chr>                            <chr>                                  <dbl>
#>  1 asteria_megalotomusquinquespino… 76269c17-78d6-420b-99…                 0.59 
#>  2 asteria_megalotomusquinquespino… 76269c17-78d6-420b-99…                 0.99 
#>  3 asteria_megalotomusquinquespino… 76269c17-78d6-420b-99…                 0.7  
#>  4 asteria_megalotomusquinquespino… 76269c17-78d6-420b-99…                 0.985
#>  5 asteria_megalotomusquinquespino… 76269c17-78d6-420b-99…                 0.59 
#>  6 asteria_megalotomusquinquespino… 76269c17-78d6-420b-99…                 0.99 
#>  7 asteria_megalotomusquinquespino… 76269c17-78d6-420b-99…                 0.7  
#>  8 asteria_megalotomusquinquespino… 76269c17-78d6-420b-99…                 0.985
#>  9 asteria_megalotomusquinquespino… 76269c17-78d6-420b-99…                 0.59 
#> 10 asteria_megalotomusquinquespino… 76269c17-78d6-420b-99…                 0.99 
#> # ℹ 1,814 more rows
#> # ℹ abbreviated name: ¹​activity_uuid_product_uuid
#> # ℹ 25 more variables: transition_risk_low_threshold <dbl>,
#> #   transition_risk_high_threshold <dbl>, transition_risk_category <chr>,
#> #   benchmark_tr_score <chr>, profile_ranking <dbl>, reduction_targets <dbl>,
#> #   country <chr>, main_activity <chr>, ep_product <chr>,
#> #   matched_activity_name <chr>, matched_reference_product <chr>, unit <chr>, …

output |> unnest_company()
#> # A tibble: 6,912 × 15
#>    companies_id           country main_activity benchmark_tr_score_avg benchmark
#>    <chr>                  <chr>   <chr>         <chr>                  <chr>    
#>  1 asteria_megalotomusqu… austria wholesaler    1.5C RPS_2030_all      all      
#>  2 asteria_megalotomusqu… austria wholesaler    1.5C RPS_2030_all      all      
#>  3 asteria_megalotomusqu… austria wholesaler    1.5C RPS_2030_all      all      
#>  4 asteria_megalotomusqu… austria wholesaler    1.5C RPS_2030_all      all      
#>  5 asteria_megalotomusqu… austria wholesaler    1.5C RPS_2030_isic_4d… isic_4di…
#>  6 asteria_megalotomusqu… austria wholesaler    1.5C RPS_2030_isic_4d… isic_4di…
#>  7 asteria_megalotomusqu… austria wholesaler    1.5C RPS_2030_isic_4d… isic_4di…
#>  8 asteria_megalotomusqu… austria wholesaler    1.5C RPS_2030_isic_4d… isic_4di…
#>  9 asteria_megalotomusqu… austria wholesaler    1.5C RPS_2030_tilt_se… tilt_sec…
#> 10 asteria_megalotomusqu… austria wholesaler    1.5C RPS_2030_tilt_se… tilt_sec…
#> # ℹ 6,902 more rows
#> # ℹ 10 more variables: emission_profile <chr>, emission_profile_share <dbl>,
#> #   profile_ranking_avg <dbl>, co2_avg <dbl>, scenario <chr>, year <chr>,
#> #   reduction_targets_avg <dbl>, transition_risk_score_avg <dbl>,
#> #   transition_risk_category <chr>, transition_risk_category_share <dbl>